+++ bycycle/tests/cyclepoints/test_extrema.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy.signal import argrelextrema
 import pytest
 
-from bycycle.cyclepoints import find_extrema, find_zerox, extrema_interpolated_phase
+from bycycle.cyclepoints.extrema import find_extrema, find_zerox
 
 # Set data path
 DATA_PATH = os.getcwd() + '/tutorials/data/'
@@ -67,28 +67,3 @@ def test_find_zerox():
     assert decays[0] < troughs[0]
     assert troughs[0] < rises[0]
     assert rises[0] < peaks[1]
-
-
-def test_extrema_interpolated_phase():
-    """Test waveform phase estimate."""
-
-    # Load signal
-    sig = np.load(DATA_PATH + 'sim_stationary.npy')
-
-    fs = 1000
-    f_range = (6, 14)
-
-    # Find peaks and troughs
-    peaks, troughs = find_extrema(sig, fs, f_range, boundary=1, first_extrema='peak')
-
-    # Find zerocrossings
-    rises, decays = find_zerox(sig, peaks, troughs)
-
-    # Compute phase
-    pha = extrema_interpolated_phase(sig, peaks, troughs, rises=rises, decays=decays)
-
-    assert len(pha) == len(sig)
-    assert np.all(np.isclose(pha[peaks], 0))
-    assert np.all(np.isclose(pha[troughs], -np.pi))
-    assert np.all(np.isclose(pha[rises], -np.pi/2))
-    assert np.all(np.isclose(pha[decays], np.pi/2))
