+++ bycycle/tests/features/test_burst.py
